Rename di back to service_container for clarity
